refactor(canonical): isolate source evidence contract
