+++ src/ledger/canonical/hncb-domestic-deposit.ts
@@ -6,9 +6,11 @@ import {
 } from "./advertised-domestic-deposit-preflight.ts";
 import {
   admitCanonicalSourceEvidence,
+  type CanonicalSourceEvidence,
+} from "./canonical-source-evidence.ts";
+import {
   commitCanonicalSourceEvidence,
   type CanonicalSourceCommitResult,
-  type CanonicalSourceEvidence,
   type CanonicalSourceStore,
 } from "./canonical-source-store.ts";
 import {
