+++ src/ledger/canonical/sinopac-domestic-deposit.check.ts
@@ -3,8 +3,8 @@ import { mkdtemp, rm } from "node:fs/promises";
 import { join } from "node:path";
 import { tmpdir } from "node:os";
 import { DatabaseSync } from "node:sqlite";
+import { admitCanonicalSourceEvidence } from "./canonical-source-evidence.ts";
 import {
-  admitCanonicalSourceEvidence,
   commitCanonicalSourceEvidence,
   createCanonicalSourceStore,
   queryCanonicalSourceCurrent,
